Abstract State definition
Move definition for enums and macro into header

Defined abstract State class for defining structure of what a State is. For this, used the entry, do, exit model for breaking down a state's behavior.

Using the model where a state's execution should be "brief" so that if it wishes to loop, it should return "Loop" instead of just running forever. This allows common functionality across states like reading in sensor values to be outside of a step. Sensor values are (optionally) read into a buffer before a state executes and then the buffer is passed by reference to the state.