Networking improvements
Fixed an issue where unexpected whitespace broke the hostapd.conf file

Switched to using bind-interfaces in the dnsmasq.conf file to get better compatibility with systemd-resolved and alternative to DNSStubListener=no in the conf for resolved.conf

Added a startup bash script for configuring the network and port and then starting the dnsmasq DHCP server and hostapd access point.

Fixed the response packet not being sent in the demo code and removed some of the logging that cluttered the console.

Began working on a readme to document the network setup. Added a screenshot for later inclusion in a readme.