Finalized state machine class heirarchy
Removed the ability for one state machine to have states with different inputs and output. Every state will have all inputs and outputs available. Not much is gained by the encapsulation to prevent a state that doesn't care about a sensor from not having access to it, but code complexity is significantly increased.

States are still declared as singletons where a new class is created for each state that is added.

States must declare how to transition to the next state from themselves. States may declare entry behavior (executed once on transition in), exit behavior (executed once on transition out), and do behavior (executed repeatedly while in the state).